fix(content-platform,content-triage): unstick the keyword-verdict triage pipeline
005: triage_submit takes p_priority smallint — the integer literal from
handle_keyword_verdict_change never resolved under named-arg matching.
006: inbox_preview_fn::text carries the arg list, so refresh_inbox_cache
built invalid dynamic SQL. content-triage/004: unqualified 'status' in
triage_submit collided with the OUT parameter. Fix migrations for
installed DBs; 001/004/002 consolidated for fresh installs.

+++ modules/content-platform/migrations/001_publish_adapter_registry.sql
@@ -214,7 +214,7 @@ BEGIN
         p_suggested_categories => NULL,
         p_suggested_from       => NULL,
         p_auto_approved_reason => NULL,
-        p_priority             => 50,
+        p_priority             => 50::smallint,
         p_metadata             => jsonb_build_object('via','keyword_verdict'),
         p_actor_id             => NULL,
         p_idempotency_key      => NULL,
